backend/requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
lxml
flask
flask_cors
python-dotenv
langgraph
langchain_community
langchain_experimental
langmem
langchain-google-genai
tavily-python
chromadb
langchainhub
faiss-cpu
sentence-transformers
pypdf
